Update tests
The original tests that focused on a specific provider didn't account for being run in an environment that had a provider set in an environment setting. This version of the tests will account for them by removing the setting(s) and it restoring it/them after each test is completed.

The settings are actually restored just before the tests so that the settings will be properly restored even if a test fails.